Frequency tracking test passed for full observations
