+++ src/UnscentedTransforms.jl
@@ -1,7 +1,8 @@
 module UnscentedTransforms
     include("_SigmaPoints.jl")
     include("_AbstractPredictor.jl")
-    #include("_StateSpaceModel.jl")
+    include("_StateSpaceModel.jl")
     
-    export SigmaParams, SigmaWeights, GaussianVar, SigmaPoints
+    export StateSpaceModel, LinearPredictor, NonlinearPredictor, GaussianVar, SigmaPoints, SigmaParams, SigmaWeights
+    export kalman_filter!, predict, predict!, update, update!
 end
