+++ test/__frequency_tracking_test.jl
@@ -15,28 +15,11 @@ Gd = ss(A,B,C,D,1.0)
 tfd = tf([1, -(1+ρ²)cos(Ω₀), ρ²], [1, -2cos(Ω₀), 1], 1.0)
 =#
 using Revise
+using UnscentedTransforms
 using StaticArrays
 using LinearAlgebra
-using PythonPlot; pygui(true)
+using Plots; plotlyjs()
 
-#include(joinpath(@__DIR__, "_StateSpaceModel.jl"))
-if !(@isdefined UnscentedKalmanFilters) 
-    includet(joinpath(@__DIR__, "UnscentedKalmanFilters.jl"))
-end
-
-using .UnscentedKalmanFilters
-
-@kwdef struct GaussianState
-    x :: Vector{Float64}
-    P :: Matrix{Float64}
-end
-
-function GaussianState(model::StateSpaceModel)
-    return GaussianState(
-        x = deepcopy(model.x),
-        P = model.PU'model.PU
-    )
-end
 
 const Δt = 0.1
 ω  = 2π/50
@@ -51,13 +34,19 @@ y[150] = 15
 #Test missing data after stabilization
 dy = [0; diff(y)]./Δt
 Y  = [dy'; y']
-Y[:,200] .= NaN
-Y[1,500] = NaN
-Y[2,501] = NaN
+#Y[:,200] .= NaN
+#Y[1,500] = NaN
+#Y[2,501] = NaN
 
-function oscillator_prediction(X, u)
+#Variances
+σ₊  = (σ+0.1)
+vsQ = [0.01*σ₊/Δt, 0.0001*σ₊, 0.1]
+vsR = [2*σ₊/Δt, σ₊]
+vsP = [10*σ₊/Δt, 10*σ₊, 10]
+
+function predictor_func(X, u)
     k = exp(X[3])
-    A = [
+    A = @SMatrix [
         0  -k   0;
         1   0   0;
         0   0   0 
@@ -65,66 +54,54 @@ function oscillator_prediction(X, u)
     return exp(A*Δt)*X #Discretize A and then predict result
 end
 
-#These methods should yield the same result
-#oscillator_observation(X, u) = X[1:2]
-oscillator_observation = ([1 0 0 ; 0 1 0], zeros(2,0))
+predictor = NonlinearPredictor(
+    f = predictor_func, 
+    Σ = Cholesky(UpperTriangular(Matrix(Diagonal(vsQ)))), 
+    θ = SigmaParams(α=1.0)
+)
 
+C = @SMatrix [1 0 0 ; 0 1 0]
+D = @SMatrix zeros(2,0)
+observer = LinearPredictor(C, D, Cholesky(UpperTriangular(Matrix(Diagonal(vsR)))))
 
-σ₊  = (σ+0.1)
-vsQ = [0.01*σ₊/Δt, 0.0001*σ₊, 0.1]
-vsR = [2*σ₊/Δt, σ₊]
-vsP = [10*σ₊/Δt, 10*σ₊, 10]
+x0 = @SVector [0, 0, log(k0)]
+state = GaussianVar(x0, Cholesky(UpperTriangular(Matrix(Diagonal(vsP)))))
 
 model = StateSpaceModel(
-    fxu = oscillator_prediction,
-    hxu = oscillator_observation,
-    x  = [0, 0, log(k0)],
-    QU = Diagonal(vsQ),
-    RU = Diagonal(vsR),
-    PU = Diagonal(vsP),
-    θ  = SigmaParams(α=1.0)
+    state = state,
+    predictor = predictor,
+    observer  = observer,
+    outlier  = 3.0
 )
-#=
-model = LinearStateSpaceModel(
-    A = zeros(3,3),
-    B = zeros(3,1),
-    C = [0 1 0],
-    Q = Hermitian(Diagonal(vsQ.^2)),
-    R = Hermitian(Diagonal(vsR.^2))
-)
-state = GaussianState(
-    x = [0, 0, log(k0)],
-    P = Hermitian(Diagonal(vsP.^2))
-)
-=#
 
-vs = [GaussianState(model)]
+vs = [state]
 
 for ii in 1:N
     kalman_filter!(model, Y[:,ii], Float64[])
-    push!(vs, GaussianState(model))
+    push!(vs, model.state)
 end
 
 
-figure()
-plot(y, ".k")
-plot([s.x[1] for s in vs[1:(end-1)]])
-plot([s.x[2] for s in vs[1:(end-1)]])
-plot([sqrt( min(5*σ, s.x[1]^2/exp(s.x[3])) + s.x[2]^2) for s in vs[1:(end-1)]]) #amplitude-equivalent energy
-legend(["measured", "velocity", "position", "energy amplitude"])
-title("Frequency Tracking Summary: UKF")
+fig = plot(y, lc=:blue, ls=:dot, label="measured")
+plot!(fig, [s.μ[1] for s in vs[1:(end-1)]], label="velocity")
+plot!(fig, [s.μ[2] for s in vs[1:(end-1)]], label="position")
+#plot([sqrt( min(5*σ, s.x[1]^2/exp(s.x[3])) + s.x[2]^2) for s in vs[1:(end-1)]]) #amplitude-equivalent energy
 
+#=
 figure()
-title("Frequency Tracking Raw State: UKF")
+labels = ["velocity", "position", "log spring"]
+
 labels = ["velocity", "position", "log spring"]
 for ii in 1:3
     subplot(3,1,ii)
     plot([s.x[ii] for s in vs[1:(end-1)]])
     ylabel(labels[ii])
 end
+title("Frequency Tracking Raw State: UKF")
+
 
 figure()
-title("Frequency Tracking Uncertainty: EKF")
+title("Frequency Tracking Uncertainty: UKF")
 labels = ["velocity", "position", "spring"]
 for ii in 1:3
     subplot(3,1,ii)
@@ -132,3 +109,4 @@ for ii in 1:3
     ylabel(labels[ii])
 end
 
+=#
